extracts testimonial form to new view
