+++ src/App.jsx
@@ -10,6 +10,7 @@ import ArticlesDashboard from './views/ArticlesDashboard'
 import ArticleCreation from './views/ArticleCreation'
 import SectionsDashboard from './views/SectionsDashboard'
 import InformationDashboard from './views/InformationDashboard'
+import TestimonialsDashboard from './views/TestimonialsDashboard'
 import InformationCreation from './views/InformationCreation'
 import LoginPage from './components/LoginPage'
 import Sidebar from './components/navigation/Sidebar'
@@ -24,12 +25,12 @@ const App = () => {
 
   useEffect(() => {
     Authentication.validateToken()
-    if (window.Cypress) {
+    //if (window.Cypress) {
       store.dispatch({
         type: 'AUTHENTICATE',
         payload: 'Johnny Cage',
       })
-    }
+    //}
   }, [authenticated])
 
   return (
@@ -51,6 +52,7 @@ const App = () => {
               path='/information/create'
               component={InformationCreation}
             />
+            <Route exact path='/testimonials' component={TestimonialsDashboard} />
           </Switch>
         </>
       ) : (
